requirements.txt
fastapi==0.68
uvicorn==0.14
python-multipart==0.0.5
sqlalchemy==1.4
psycopg2==2.9.1
pydantic==1.8.2
pandas==1.3.1
scikit-learn==0.24.2
scipy==1.7.0
numpy==1.21.1
requests==2.26.0
joblib==1.0.1
pyyaml==5.4.1